Integration test build will only be a failure if 40%+ tests fail.
We don't expect all tests to pass yet because of awaiting-content, 302s on GOV.UK and other things, but we know that more than 60% of tests pass at the moment, so we want to know that roughly, we've not made the situation worse. This figure can be put up as the percentage of tests we think should pass goes up.