Back out changes introduced in 7e899baa
The problem was that the 410 error page was being served by the same location
block as was throwing the 410. Nginx detected this and served its default 410
page to avoid an infinite loop.

A slightly icky (but nonetheless more functional) workaround is included here.
Explicitly specify empty location directives for each of the error pages.