Do not generate config for mappings for the wrong domain
Config is created for each domain in an input mappings.csv.
So if the CSV has mappings for old URLs starting www.direct.gov.uk and direct.gov.uk, then two separate config files are created.

This is a problem when, for example, config of an extant domain exists in another input.csv. For example if there was www.businesslink.gov.uk config in the www.ukwelcomes.businesslink.gov.uk input file, then when the ukwelcomes config was generated, that businesslink config would overwrite the previously generated businesslink config.

This solution creates a ${site_incorrect}.csv in dist for any mapping with the wrong domain and does not create config for those. They will need to be handled manually.