No longer required to be in an if block
The Businesslink config only generated a map key if there was a query string, and if there wasn't a query string, called back to the generic location config. There is now no need for this if block as we are only calling the Businesslink config generator if there is a query string.