Ensure we include document_mappings attachments
Previously we were using the "Whole Tag" field to show whether to
include assets or not in the final output, but the document_mappings
file is now generating a list which sets the "State" field to "Closed".

Therefore we needed to change this script to include rows with a
"Closed" value for "State".

This will require a remunging of all recent departments to see the
effect.

See: https://www.pivotaltracker.com/story/show/44453265