Split out munge process into fetch + munge
We now use the MappingFetcher ruby library in two different ways: as a
fetcher which generates a temporary csv file (fetch.<dept>.csv) and as a
generic munge script (`./munge/munge.rb`).

This allows us to re-run the munge locally without needing to go to the
network for the spreadsheets and re-combine the munged file each time.
In order to do a complete re-munge from the source spreadsheets you will
need to remove fetch.<dept>.csv and document_mappings.csv from the
redirector folder.

This is a precursor to splitting the fetch and munge code completely from the
MappingFetcher.

`generate-redirects.sh` has been updated to show the pipeline and can be
run unchanged.