Correct erroneous "popular" tests
These were expecting an outcome based on the behaviour of redirector, rather
than the intent of the mappings.

The reason redirector returned what it did was because it was falling back to
matching the querystring only. This meant that it would return the new URL from
an unrelated mapping.

These tests will now fail against redirector, but pass against bouncer.

The deleted test had to go because there is no mapping to provide the behaviour
but adding a mapping would break the redirector (smokey?) build. The correct
redirect URL would be:
https://www.gov.uk/government/organisations/department-for-work-pensions/series/family-resources-survey--2