Change %2f => %2F
It's worth noting that these two are considered to both represent a forward
slash. This change is to allow bouncer to use these mappings.

Bouncer will canonicalise a request with %2f or %2F to %2F. It then looks for
the mapping with the capitalised version.

Because these mappings were lower case, bouncer wouldn't find them for the
canonicalised URL. This meant that smoke tests failed.

Hopefully, this change will make no difference to redirector.