Remove special global rule for treasury_cdn
This site has a special global redirect rule, which for Redirector
was in its site YAML. This rule is already hard-coded into Bouncer,
and the Transition import already ignores the global status for
this site as a special case, so we don't need to keep this rule in
this file any more. If this site needs to be moved back to Redirector
in future, this change will need to be reverted.