Add all known EA querystrings for Flood Warnings
These are all of the EA querystrings we currently know about in use on Flood
Warnings.

We don't intend to write mappings for these URLs, we will have a rule. However,
because mappings are evaluated before rules, we want to make sure that if
someone did create rules for these URLs that they wouldn't be canonicalised
such that it would map all of these URLs.

Basically, for EA Flood Warnings, we always want the rule. Adding these
significant querystring paramters is a safeguard.