procurement.cab... => aka-procurement.cab...
procurement.cabinetoffice.gov.uk cannot be CNAMEd, so we need to host it at an
IP and redirect to aka-procurement.cabinetoffice.gov.uk which can be CNAMED.

Whilst procurement.cabinetoffice.gov.uk isn't at the root of the DNS zone, it
does have MX records for receiving mail. The DNS specification prevents you
from having both a CNAME and MX record for a domain.