Add a rake task to validate that hosts appear only once
A host should never appear on multiple sites at once. We want to be
able to reassign hosts to other sites in Transition's orgs_sites_hosts
import, and this task makes that a little safer. The task is included
in the CI build so that we can see clearly if any duplicate hosts are
present across all site yaml files.