titles no longer required in site yaml
Transition and Bouncer currently ignore site titles, although once upon a time
they were used in the import process to relate them to their owning
organisation. We now do this with the whitehall_slug.

Instead, let's have titles only when it should be different from the
organisation title - for example when the site is about something other than
the organisation. We'll change Transition and Bouncer to use the site title or
fallback to the organisation title.

Where there is an obvious title to use instead of the organisation title (or
the title was already different), I've done so.