Reinstate old Worldpay callback proxy_pass
We didn't realise that this is a POST request which therefore can't be
redirected. Because they couldn't be redirected, the outcome of licencing
application payments via Worldpay isn't being recorded.

We are hoping to get Worldpay to POST to the new URL instead, but for now we
need to not break licence applications.