Add tlds as aliases to existing sites
The "tld" files exist so that we can redirect domains at the root of their zone.
However, they aren't well supported, add a layer of complexity and aren't a
long-term solution.

This change is a step in the process of replacing that.

Deeper explanation:
Domains at the root of their zone cannot be CNAMEd, as is our usual practice,
they can have an A record which points to an IP. However, because our CDN hasn't
supported stable IPs, we haven't been able point these domains at our CDN.

Instead, we get transitioning organisations to create an A record which points
at a box with a stable IP which has the "tld" nginx configuration on it. Each
of these tld files redirects to the www. subdomain of the domain, which is
CNAMEd to our CDN.

Now, though, we have stable IPs for our CDN. This means we can get transitioning
organisations to use those for A records/domains.