Ignore ordering in hosts_to_site_abbrs
test_hosts_to_site_abbrs_when_a_host_appears_twice has started to fail
intermittently in CI due to differences in ordering between the actual
and expected output, but only when it is run on the new ci-slave-5
(which is Trusty, not Precise):

expected: {"one.local"=>["one"], "alias1.one.local"=>["one"], "alias2.one.local"=>["one"], "two.local"=>["one", "two"]}
actual:   {"two.local"=>["two", "one"], "one.local"=>["one"], "alias1.one.local"=>["one"], "alias2.one.local"=>["one"]}

We don't care about the ordering here, in fact, so this commit uses
sets as the values in the output rather than arrays. With this change,
the test considers the output to be equal no matter which order the
files are read.

(We don't know why files are now being read in a different order,
but it has revealed an assumption in our code which was unnecessary.)