Update highways homepage and whitehall_slug
Highways Agency has been replaced by Highways England, and the homepage
for this site needs to redirect to the new organisation homepage on GOV.UK.
The whitehall_slug also needs to be updated so that users belonging to the
new organisation can edit mappings for the site.

The homepage furl is already redirecting to the new page so no action is
needed there.

There are currently 2 other sites which are owned by highways-agency
(ha_standardsforhighways and ha_trafficengland); we're waiting to hear whether
those also need the same config changes, but are going ahead with this one
now.