Give access to Highways England/Highways Agency sites to both orgs
Highways Agency has become Highways England. This has been modeled as the old
organisation closing and a new one being created. In Signon, some users have
been reassigned to the new organisation, some haven't. The `highways.yml` site
is associated with the new organisation, but other sites are associated with
the old one. This means that users can't access all the sites they need to.

If we make all of their sites accessible to members of (and displayed under)
both organisations, it doesn't matter which organisation the user is associated
with.