Fix ukti_arabspring site naming
The filename for this site didn't match the site name in the file (which was
a different site's name) so the www.arabspring.ukti.gov.uk host was wrongly
imported as an alias of ukti_lps. Both sites are global 410 so the behaviour
will not change, although the sites have different TNA timestamps.

This commit also renames the site to make it closer to the hostname.