Rename site file to match name in file
The site has already been imported using the name in the file, so this change
to the filename won't cause any changes on import. Using the name ending in 1
doesn't look great, but it's OK in this case because the main host for the
site is careersadvice-findacourse1.direct.gov.uk and that host is used as the
displayed domain for the site.