Fix site name for bis_nmodoitonline
This site name didn't match the filename but was the same as another site's
name, so its hosts have been imported as aliases of the other site. This
change to the name in the file will create a new site and reassign the hosts
to the new site on import.