More helpful error message for `sitemap.sh`
This script uses wget.

Before the changes in this commit, if the script's user supplied an
un-crawlable URL, they'd see an unrelated error later on in the script.

Since wget won't blow up and propagate an exception to the user, the script
needs to check wget's exit status and exit early itself.