Tool for checking bad URLs
This tool provides almost the exact opposite function to `strip_mappings.sh`.
Instead of printing the "good" subset of input URLs, it prints the "bad" ones.

This was intended to be a throwaway script just to double check our use of
`strip_mappings.sh` and to verify the entries in `strip_list.txt`.

Considering how much use we ended up getting out of it, it seems like it might
be useful for someone in future.