Switch host and alias around for dclg_planningportal
The www domain is the main live one for this site, and has the aka domain set
up. We need the www domain to be the default one for the site so that the
links to the side-by-side browser work correctly.

The import isn't able to change the default host for a site, so I have already
manually switched the hostnames around (including the aka ones). This change
to the site YAML file is only for consistency; importing it without the manual
change would have no effect.