Add hmrc_inrev site as global redirect
This is a collection of very old HMRC domains which have very little or no
history in TNA. I don't know if they were ever really aliases of the same site
originally since there's so little evidence of them serving anything, but
we've been asked to configure them all as global redirects so I've put them
all in this one site. If we find a reason for them to be separate sites later
on we can split some domains out.

This has a fake timestamp because there aren't any useful crawls to set for
these domains - since they're globally redirecting the TNA links won't be
shown to users anyway.