Add hmrc_paymastergeneral site as a global redirect
This timestamp is for the latest crawl which doesn't redirect to another
domain (there's only one later one, which redirects to hmrc.gov.uk).

That TNA crawl shows that in this case OPG refers to the Office of HM Paymaster
General and not the Office of the Public Guardian. The former was incorporated
into the Government Banking Service in 2008, but has a closed organisation
page on GOV.UK, so I've used that as the destination for the global redirect.

Since the old org has closed, there's no point in adding it as an extra
organisation because it won't have any users in Signon or Transition.