Split PHE obesitywm and foodwm sites
We originally configured these domains as a single site because they appeared
to be identical in the archives and the timestamps seemed to work
interchangeably.

Now http://webarchive.nationalarchives.gov.uk/20150116162451/http://www.foodwm.org.uk/
returns a 401 error, so we're splitting them into two separate sites to use
the official timestamp for each one.

The import:revert:sites rake task will need to be run in Transition for the
already-imported phe_foodobesitywm site when this is merged so that the
database remains consistent when this change is imported.