Fix site abbr for gds_alphagov_licensify-admin-preview
A separate site hasn't been created for this config file because it has the
same abbr as another site. We haven't yet switched from Preview to Integration
so the uncreated site isn't live yet.

This was found by the new validation which checks that a site's abbr matches
its filename.