Check that YAML files are only in the expected directories
We've occasionally merged new site config files which were in the root of the
repo instead of in data/transition-sites/ and so weren't imported. Check as
part of the build that *.yml files are only present in data/transition-sites/
and in the test fixtures directory. We also need to exclude the vendor
directory so that installed gems don't trigger this check.

See https://github.com/alphagov/transition-config/pull/1146 for a recent
example of this.