Use filename instead of site.abbr as key for missing fields
If the site field is missing from a file then site.abbr will be nil, which
makes it unhelpful to use as the key in the hash of missing fields. Use the
filename instead so that it's clearer where the problem is, and to prevent
multiple files missing the site field overwriting each other in the hash.