Update homepage and homepage_title for www.localdirect.gov.uk
This site is archived at this domain, but Local Digital Coalition are going to
continue to run and update the site at a non-government domain, so we need to
link to that instead of the DCLG homepage from the archive pages.

This commit also removes the css because we don't think we need it here.