Alter the handling of two UKVI related domains
These domains are currently handled through Dyn (dyn.com), and the
configuration for each domain is to redirect all requests to a single
page (different page for each domain).

The configuration here for these domains is to handle them as aliases
for www.ukba.homeoffice.gov.uk, which has a set of mappings in the
transition app to give different redirects based on the request path.

To make the behaviour of transition/bouncer match that currently setup
through Dyn, split the handling of these two domains out, and
configure the respective global redirects.

This will mean that should the traffic be sent to bouncer, the
behaviour stay the same. This will enable removing the configuration
from Dyn.