Add GDS Legislation global redirect
The domain legislation.data.gov.uk is currently redirected using a
legacy Akamai service, which we would like to discontinue. A possible
consequence of this existing redirect is that this site is not present
in the National Archive, except for a couple of specific pages upon
which the tna_timestamp value is based.