Add civil mediation site config
The new homepage domain will need adding to the redirection whitelist as it's
not a *.gov.uk, *.nhs.uk or *.mod.uk subdomain.

The TNA timestamp I've used is the most recent one - it won't ever be seen by
users anyway since this site is being configured as a global redirect.