Redirect all UK Gov in Wales domains to GOV.UK
The Office of the Secretary of State for Wales now uses
ukgovwales.gov.uk as their primary domain.

They are looking to reach citizens in Wales using social media
advertising, and Facebook requires that their email domain matches that
of their primary website. Given that, they need ukgovwales.gov.uk to
redirect to their homepage on GOV.UK to allow these campaigns to direct
users to the right place.