Remove www.siriusprogramme.com alias
`dig siriusprogramme.com`, `dig aka.siriusprogramme.com` and `dig
www.siriusprogramme.com` return NXDOMAIN, and there’s no WHOIS record for
siriusprogramme.com anymore either.
There’s no domain there for us to attach a TXT record to, which is required to
manually verify domains to use them in Fastly.

The addition of this domain is blocking govuk-fastly terraform apply as this
project fetches a list of domains from Transition, here:
https://transition.publishing.service.gov.uk/hosts.json

We are removing it from the config to un-block TF apply.  The records will need
to be separately updated in Transition as the script only adds sites.